Move safe_statements_spec.rb helpers into shared location
This will enable us to more easily split this file up.
