+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ require "bundler/setup"
 require "pg_ha_migrations"
 require "db-query-matchers"
 require "pry"
+require "test_helpers"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
